+++ lib/src/customized_dialog.dart
@@ -13,7 +13,8 @@ enum DialogTipType {
 class ProsteDialog extends StatefulWidget {
   ProsteDialog({
     Key? key,
-    this.insetPadding = const EdgeInsets.symmetric(horizontal: 40.0, vertical: 24.0),
+    this.insetPadding =
+        const EdgeInsets.symmetric(horizontal: 40.0, vertical: 24.0),
     this.dialogRadius = 5,
     this.type = DialogTipType.success,
     this.title,
@@ -33,9 +34,9 @@ class ProsteDialog extends StatefulWidget {
     this.backgroundColor = Colors.white,
     this.elevation = 0,
     this.shadowColor,
-    this.btnInARow = true,
+    this.btnsInARow = true,
     this.btnPadding,
-    this.columnConfirmBtnMargin,
+    this.confirmButtonMargin,
   }) : super(key: key);
 
   /// dialog与屏幕之间的间距
@@ -99,19 +100,20 @@ class ProsteDialog extends StatefulWidget {
   final Color? shadowColor;
 
   /// 按钮是否放在同一行
-  final bool btnInARow;
+  final bool btnsInARow;
 
   /// 按钮的padding
   final EdgeInsets? btnPadding;
 
   /// 不是同一行是确认按钮的margin
-  final EdgeInsets? columnConfirmBtnMargin;
+  final EdgeInsets? confirmButtonMargin;
 
   @override
   _ProsteDialogState createState() => _ProsteDialogState();
 }
 
-class _ProsteDialogState extends State<ProsteDialog> with TickerProviderStateMixin {
+class _ProsteDialogState extends State<ProsteDialog>
+    with TickerProviderStateMixin {
   /// 演示关闭控制器
   Timer? _time;
 
@@ -145,20 +147,33 @@ class _ProsteDialogState extends State<ProsteDialog> with TickerProviderStateMix
       _time = Timer(widget.duration!, () => Navigator.pop(context));
     }
     // 判断动画类型
-    _controller = AnimationController(vsync: this, duration: Duration(seconds: 1))..forward();
-    _scaleController = AnimationController(vsync: this, duration: Duration(milliseconds: 300))..forward();
+    _controller =
+        AnimationController(vsync: this, duration: Duration(seconds: 1))
+          ..forward();
+    _scaleController =
+        AnimationController(vsync: this, duration: Duration(milliseconds: 300))
+          ..forward();
     switch (widget.type) {
       case DialogTipType.success:
-        _scaleAnimation = Tween<double>(begin: .3, end: 1).animate(_scaleController);
-        _rotationAnimation = Tween<double>(begin: pi / 2, end: 0).chain(CurveTween(curve: Curves.bounceOut)).animate(_controller);
+        _scaleAnimation =
+            Tween<double>(begin: .3, end: 1).animate(_scaleController);
+        _rotationAnimation = Tween<double>(begin: pi / 2, end: 0)
+            .chain(CurveTween(curve: Curves.bounceOut))
+            .animate(_controller);
         break;
       case DialogTipType.warn:
-        _scaleAnimation = Tween<double>(begin: 1.5, end: 1).animate(_scaleController);
-        _translationAnimation = Tween<double>(begin: -25, end: 0).chain(CurveTween(curve: Curves.bounceOut)).animate(_controller);
+        _scaleAnimation =
+            Tween<double>(begin: 1.5, end: 1).animate(_scaleController);
+        _translationAnimation = Tween<double>(begin: -25, end: 0)
+            .chain(CurveTween(curve: Curves.bounceOut))
+            .animate(_controller);
         break;
       case DialogTipType.error:
-        _scaleAnimation = Tween<double>(begin: .3, end: 1).animate(_scaleController);
-        _rotationAnimation = Tween<double>(begin: pi / 2, end: 0).chain(CurveTween(curve: Curves.elasticOut)).animate(_controller);
+        _scaleAnimation =
+            Tween<double>(begin: .3, end: 1).animate(_scaleController);
+        _rotationAnimation = Tween<double>(begin: pi / 2, end: 0)
+            .chain(CurveTween(curve: Curves.elasticOut))
+            .animate(_controller);
         break;
     }
   }
@@ -207,7 +222,8 @@ class _ProsteDialogState extends State<ProsteDialog> with TickerProviderStateMix
               return Transform.scale(
                 scale: _scaleAnimation!.value,
                 child: Transform(
-                  transform: Matrix4.translationValues(0, _translationAnimation!.value, 0),
+                  transform: Matrix4.translationValues(
+                      0, _translationAnimation!.value, 0),
                   child: child,
                 ),
               );
@@ -253,8 +269,10 @@ class _ProsteDialogState extends State<ProsteDialog> with TickerProviderStateMix
       contentPadding: widget.contentPadding,
       confirmButtonText: widget.confirmButtonText,
       cancelButtonText: widget.cancelButtonText,
-      showConfirmButton: widget.duration == null ? widget.showConfirmButton : false,
-      showCancelButton: widget.duration == null ? widget.showCancelButton : false,
+      showConfirmButton:
+          widget.duration == null ? widget.showConfirmButton : false,
+      showCancelButton:
+          widget.duration == null ? widget.showCancelButton : false,
       buttonRadius: widget.buttonRadius,
       confirmButtonColor: widget.confirmButtonColor,
       cancelButtonColor: widget.cancelButtonColor,
@@ -263,9 +281,9 @@ class _ProsteDialogState extends State<ProsteDialog> with TickerProviderStateMix
       backgroundColor: widget.backgroundColor,
       elevation: widget.elevation,
       shadowColor: widget.shadowColor,
-      btnInARow: widget.btnInARow,
+      btnInARow: widget.btnsInARow,
       btnPadding: widget.btnPadding,
-      columnConfirmBtnMargin: widget.columnConfirmBtnMargin,
+      confirmButtonMargin: widget.confirmButtonMargin,
     );
   }
 }