+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0]
+
+* add button vertical layout
+
+* add btnPadding
+
+* add confirmButtonMargin, and it is only butsInARow is false is it valid
+
 ## [1.2.0]
 
 * add backgroundColor and shadow